refactor: move synfig::filesystem::Path class to its own source files
