+++ synfig-core/src/synfig/filesystem_path.cpp
@@ -1,0 +1,205 @@
+/* === S Y N F I G ========================================================= */
+/*!	\file filesystem_path.cpp
+**	\brief class Path that handles Path components and conversions
+**
+**	\legal
+**	Copyright (c) 2002-2005 Robert B. Quattlebaum Jr., Adrian Bentley
+**	......... ... 2022 Synfig Contributors
+**
+**	This file is part of Synfig.
+**
+**	Synfig is free software: you can redistribute it and/or modify
+**	it under the terms of the GNU General Public License as published by
+**	the Free Software Foundation, either version 2 of the License, or
+**	(at your option) any later version.
+**
+**	Synfig is distributed in the hope that it will be useful,
+**	but WITHOUT ANY WARRANTY; without even the implied warranty of
+**	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+**	GNU General Public License for more details.
+**
+**	You should have received a copy of the GNU General Public License
+**	along with Synfig.  If not, see <https://www.gnu.org/licenses/>.
+**	\endlegal
+*/
+/* ========================================================================= */
+
+/* === H E A D E R S ======================================================= */
+
+#ifdef USING_PCH
+# include "pch.h"
+#else
+# ifdef HAVE_CONFIG_H
+#  include <config.h>
+# endif
+
+# include "filesystem_path.h"
+
+# ifdef _WIN32
+#  include <codecvt>
+#  include <locale>
+
+#  include "general.h" // synfig::error(...)
+# endif
+
+#endif
+
+/* === U S I N G =========================================================== */
+
+using namespace synfig;
+
+/* === M A C R O S ========================================================= */
+
+/* === G L O B A L S ======================================================= */
+
+/* === P R O C E D U R E S ================================================= */
+
+/* === M E T H O D S ======================================================= */
+
+filesystem::Path::Path()
+{
+}
+
+filesystem::Path::Path(const std::string& path)
+	: path_(path), native_path_dirty_(true)
+{
+}
+
+const filesystem::Path::value_type*
+filesystem::Path::c_str() const noexcept
+{
+	return native().c_str();
+}
+
+const filesystem::Path::string_type&
+filesystem::Path::native() const noexcept
+{
+	return native_path_;
+}
+
+const std::string&
+filesystem::Path::u8string() const
+{
+	return path_;
+}
+
+filesystem::Path
+filesystem::Path::filename() const
+{
+	auto filename_pos = get_filename_pos();
+
+	if (filename_pos == std::string::npos)
+		return Path();
+
+	return path_.substr(filename_pos);
+}
+
+filesystem::Path
+filesystem::Path::stem() const
+{
+	auto filename_pos = get_filename_pos();
+	if (filename_pos == std::string::npos)
+		return Path();
+
+	auto extension_pos = get_extension_pos();
+
+	auto stem_length = extension_pos == std::string::npos ? extension_pos : extension_pos - filename_pos;
+
+	return path_.substr(filename_pos, stem_length);
+}
+
+filesystem::Path
+filesystem::Path::extension() const
+{
+	auto extension_pos = get_extension_pos();
+
+	if (extension_pos == std::string::npos)
+		return Path();
+
+	return path_.substr(extension_pos);
+}
+
+bool
+filesystem::Path::empty() const noexcept
+{
+	return path_.empty();
+}
+
+bool
+filesystem::Path::has_filename() const
+{
+	return get_filename_pos() != std::string::npos;
+}
+
+bool
+filesystem::Path::has_stem() const
+{
+	auto filename_pos = get_filename_pos();
+	auto extension_pos = get_extension_pos();
+	return filename_pos != std::string::npos && (extension_pos == std::string::npos || filename_pos < extension_pos);
+}
+
+bool
+filesystem::Path::has_extension() const
+{
+	return get_extension_pos() != std::string::npos;
+}
+
+std::size_t
+filesystem::Path::get_filename_pos() const
+{
+	if (path_.empty())
+		return std::string::npos;
+	auto separator_pos = path_.find_last_of("/\\");
+	if (separator_pos == std::string::npos)
+		return 0;
+	if (separator_pos + 1 == path_.size())
+		return std::string::npos;
+	return separator_pos + 1;
+}
+
+std::size_t
+filesystem::Path::get_extension_pos() const
+{
+	auto dot_pos = path_.rfind('.');
+	// no dot char
+	if (dot_pos == std::string::npos)
+		return std::string::npos;
+
+	auto filename_pos = get_filename_pos();
+	// no filename? no extension then
+	if (filename_pos == std::string::npos)
+		return std::string::npos;
+
+	// last dot  char was found before filename? not an extension separator then
+	if (filename_pos > dot_pos)
+		return std::string::npos;
+
+	// path is hidden file (.foo) or special dot file
+	if (filename_pos == dot_pos)
+		return std::string::npos;
+
+	// path is special dot-dot (..)
+	if (path_.size() - filename_pos == 2 && path_.compare(filename_pos, 2, "..") == 0)
+		return std::string::npos;
+
+	return dot_pos;
+}
+
+filesystem::Path::string_type
+filesystem::Path::utf8_to_native(const std::string& utf8)
+{
+#ifdef _WIN32
+	// Windows uses UTF-16 for filenames, so we need to convert it from UTF-8.
+	try {
+		std::wstring_convert<std::codecvt_utf8_utf16<wchar_t>> wcu8;
+		return wcu8.from_bytes(utf8);
+	} catch (const std::range_error& exception) {
+		synfig::error("Failed to convert UTF-8 string (%s)", utf8.c_str());
+		throw;
+	}
+#else
+	// For other OS, it's the file name as it is
+	return utf8;
+#endif
+}
